modified HTML to make the toip-level index.html validate
This required a bit of non-trivial alteration, but the result looks
about the same (perhaps better), and there are fewer tags.  Some
broken things were fixed.  And I changed the doctype to xhtml 1.0
transitional (i.e. not strict) so that iframe tags can be used (strict
does not support them).